requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.32.0
kiteconnect~=5.0.0
apscheduler~=3.10.0
sqlalchemy~=2.0.0
alembic~=1.14.0
psycopg2-binary~=2.9.0
pyyaml~=6.0
pandas~=2.2.0
numpy~=2.1.0
typer~=0.12.0
httpx~=0.27.0
yfinance~=0.2.40
pyarrow~=17.0.0
xlsxwriter~=3.2.0
